+++ packages/backend/lambda/Makefile
@@ -1,4 +1,4 @@
 build:
 	mkdir -p dist
 	go get ./...
-	go build -o dist/hello ./...
+	go build -o functions/hello ./...
