+++ packages/backend/lambda/functions/hello.go
@@ -1,0 +1,17 @@
+package main
+
+import (
+	"github.com/aws/aws-lambda-go/events"
+	"github.com/aws/aws-lambda-go/lambda"
+)
+
+func handler(request events.APIGatewayProxyRequest) (*events.APIGatewayProxyResponse, error) {
+	return &events.APIGatewayProxyResponse{
+		StatusCode: 200,
+		Body:       "Go, Lambda",
+	}, nil
+}
+
+func main() {
+	lambda.Start(handler)
+}
