+++ packages/backend/lambda/netlify.toml
@@ -1,7 +1,6 @@
 [build]
   command = "make build"
   functions = "src"
-  publish = "dist"
 [build.environment]
   # Change this path with the path to your repository
   GO_IMPORT_PATH = "https://github.com/chaos-monkeys/git-chaos/"
